refactor(domain): organise directory structure
