+++ domain/src/main/java/com/my/kizzy/domain/model/logs/LogEvent.kt
@@ -2,7 +2,7 @@
  *
  *  ******************************************************************
  *  *  * Copyright (C) 2022
- *  *  * LogLevel.kt is part of Kizzy
+ *  *  * LogEvent.kt is part of Kizzy
  *  *  *  and can not be copied and/or distributed without the express
  *  *  * permission of yzziK(Vaibhav)
  *  *  *****************************************************************
@@ -10,8 +10,11 @@
  *
  */
 
-package com.my.kizzy.domain.model
+package com.my.kizzy.domain.model.logs
 
-enum class LogLevel {
-    INFO,DEBUG,WARN,ERROR
-}
+data class LogEvent(
+    var level: LogLevel,
+    var tag: String,
+    var text: String,
+    val createdAt: Long
+)
