+++ domain/src/main/java/com/my/kizzy/domain/model/rpc/RpcConfig.kt
@@ -1,4 +1,17 @@
-package com.my.kizzy.domain.model
+/*
+ *
+ *  ******************************************************************
+ *  *  * Copyright (C) 2022
+ *  *  * RpcConfig.kt is part of Kizzy
+ *  *  *  and can not be copied and/or distributed without the express
+ *  *  * permission of yzziK(Vaibhav)
+ *  *  *****************************************************************
+ *
+ *
+ */
+
+package com.my.kizzy.domain.model.rpc
+
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
