+++ domain/src/main/java/com/my/kizzy/domain/model/user/Badge.kt
@@ -2,7 +2,7 @@
  *
  *  ******************************************************************
  *  *  * Copyright (C) 2022
- *  *  * RpcButtons.kt is part of Kizzy
+ *  *  * Badge.kt is part of Kizzy
  *  *  *  and can not be copied and/or distributed without the express
  *  *  * permission of yzziK(Vaibhav)
  *  *  *****************************************************************
@@ -10,19 +10,15 @@
  *
  */
 
-package com.my.kizzy.domain.model
+package com.my.kizzy.domain.model.user
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class RpcButtons(
-    @SerialName("button1")
-    val button1: String = "",
-    @SerialName("button2")
-    val button2: String = "",
-    @SerialName("button1Url")
-    val button1Url: String = "",
-    @SerialName("button2Url")
-    val button2Url: String = ""
+data class Badge(
+    @SerialName("icon")
+    val icon: String,
+    @SerialName("name")
+    val name: String
 )
