+++ feature_custom_rpc/src/main/java/com/my/kizzy/feature_custom_rpc/components/sheet/SaveConfigDialog.kt
@@ -24,7 +24,7 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import com.blankj.utilcode.util.FileIOUtils
-import com.my.kizzy.domain.model.RpcConfig
+import com.my.kizzy.domain.model.rpc.RpcConfig
 import com.my.kizzy.resources.R
 
 @Composable
