+++ feature_logs/src/main/java/com/my/kizzy/feature_logs/LogScreen.kt
@@ -38,7 +38,7 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.text.withStyle
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.my.kizzy.domain.model.LogEvent
+import com.my.kizzy.domain.model.logs.LogEvent
 import com.my.kizzy.preference.Prefs
 import com.my.kizzy.ui.components.KSwitch
 import com.my.kizzy.ui.components.SearchBar
