+++ feature_logs/src/main/java/com/my/kizzy/feature_logs/LogsViewModel.kt
@@ -15,7 +15,7 @@ package com.my.kizzy.feature_logs
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.toMutableStateList
 import androidx.lifecycle.ViewModel
-import com.my.kizzy.domain.model.LogEvent
+import com.my.kizzy.domain.model.logs.LogEvent
 import com.my.kizzy.feature_logs.LoggerProvider.logger
 import com.my.kizzy.preference.Prefs
