+++ feature_profile/src/main/java/com/my/kizzy/feature_profile/ui/user/UserViewModel.kt
@@ -18,7 +18,7 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.my.kizzy.domain.model.Resource
-import com.my.kizzy.domain.model.User
+import com.my.kizzy.domain.model.user.User
 import com.my.kizzy.domain.use_case.get_user.GetUserUseCase
 import com.my.kizzy.preference.Prefs
 import dagger.hilt.android.lifecycle.HiltViewModel
