+++ feature_settings/src/main/java/com/my/kizzy/feature_settings/Settings.kt
@@ -31,7 +31,7 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.my.kizzy.domain.model.User
+import com.my.kizzy.domain.model.user.User
 import com.my.kizzy.resources.R
 import com.my.kizzy.ui.components.Subtitle
 import com.my.kizzy.ui.components.chips
@@ -186,7 +186,6 @@ fun RequestQsTile(
 }
 
 
-@OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun SettingsItemCard(
     title: String,
